Collectively track orders split up into separate orders by Amazon (#115)
* Collectively track orders split up into separate orders by Amazon

This has the effect of merging back together orders that Amazon has "split up"
into several orders, which generally is not what you want because buying groups
often reimburse different items in combos differently than how Amazon bills
them, and if the sub-orders aren't merged together, then many will appear to be
reimbursed below cost and others above cost.

This commit blows away the old expected_costs cache (because it did not contain
the email ID information which is necessary to be able to merge sub-orders
together). Thus, the first run after upgrading to this change is going to take
a long time because it will need to get all of the shipments info again from
scratch. Fortunately that won't take as long as it used to since this PR also
changes order email fetching behavior to reuse the same login (thanks Gus!).

This fixes #114.