Allow manual_input.py to add order #s to an existing tracking #
This is useful in the case of packages being shipped that contain multiple
orders that are combined together, because the "has shipped" email notification
does NOT contain all order #s for personal accounts. The only solution is to
painstakingly go through your list of all orders and manually enter the order #s
that are missing from the reconciliation sheet, along with the tracking #s that
they correspond to. Then run reconcile.py again and everything should be there.

This change also does some refactoring to the tracking_output module, in
particular making config an instance variable instead of requiring it on each
function call.