Add a get_tracking.py --seen scanning mode
This will go through all of the *seen* shipped emails in the lookback period
(which defaults to 45 days) instead of all of the unseen ones. This is a much
larger number of emails generally, and is useful for catching shipping
notifications and tracking #s that may have been missed on the first pass for
whatever reason, e.g. because an email was accidentally manually opened prior to
being ingested by get_tracking, or Amazon has changed the tracking # for the
order without sending a new notification (which happens on rare occasions).

So it's thus a good idea to run this new mode every so often, probably every
couple weeks or so, to ensure that no shipped orders have slipped through the
cracks and were never picked up. Yes, this takes a long time; run it over night.

Subsequent PRs will add smarter functionality to this mode, including an errored
tracking # accumulator to be output at the end, better retry functionality, and
some kind of progress indicator.

This partially implements #118.