Add retrying to retrieval of tracking #s
This should be able to cope with the dreaded timeout errors that ruin an entire
run of get_tracking_numbers.py, and might also address CAPTCHAs by exponentially
backing off.

Since there's now retrying, we can afford to be less conservative on our waits,
so we go faster as well.