Limit importing of manual override status from Sheet
We only want to import the manual override status from the Google Sheet if the
cluster is still exactly the same (no merges, no added orders or trackings). If
the cluster is no longer the same then whatever the user set on the manual
override checkbox is out of date and shouldn't be imported.

This also makes the Notes merging behavior smarter by checking if strings are
blank, not just empty, so you won't get long strings of commas anymore. It also
switches over to semi-colons, which is better because commas can often be found
in the text of an individual note.

This fixes #131.