Add handling of additional cancellation email formats (#148)
Amazon sends emails in a different format if you requested the cancellation vs
if they cancelled on you for lack of payment. This code handles all formats
known to me. Unfortunately, not all these formats include the item quantity
cancelled, so "??" is used to mark indeterminate quantity.

To regenerate all cancellation data, delete `output/cancellations.pickle` and
the copy on Google Drive and then run this latest version.